refactor: photographer creation

+++ src/main/java/br/edu/ifpb/photobean/model/Comment.java
@@ -10,9 +10,10 @@ import java.time.LocalDateTime;
 @Getter
 @Setter
 public class Comment {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     private String commentText;
